Add a fail handler to the smoke tests
Also dry up all end to end tests making them reuse more bits like
- fail handlers and common funcs used in them (like printing pod logs)
- env helpers
