+++ tests/e2e/spaces_test.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"sync"
 
-	"code.cloudfoundry.org/korifi/tests/e2e/helpers"
+	"code.cloudfoundry.org/korifi/tests/helpers"
 
 	"github.com/go-resty/resty/v2"
 	"github.com/google/uuid"
